requirements.txt
django
djangorestframework
pyjwt
bcrypt
requests